+++ Tests/Documentation/ObservableDocumentationTest.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using Optional;
 using UnityEngine;
 using Yaga.Utils;
 
@@ -30,8 +31,8 @@ namespace Yaga.Test.Documentation
         public void BindingChainsExample()
         {
             var amount = new Observable<int>(100);
-            var itemInfo = Observable.Bind(amount, count => $"Current amount is {count}");
-            var disposable = itemInfo.Add(info => Assert.AreEqual("Current amount is 5", info));
+            var itemInfo = amount.Select(count => $"Current amount is {count}");
+            var disposable = itemInfo.Subscribe(info => Assert.AreEqual("Current amount is 5", info));
             amount.Data = 5;
 
             disposable.Dispose();
@@ -43,11 +44,11 @@ namespace Yaga.Test.Documentation
         {
             var amount = new Observable<int>(100);
             var itemName = new Observable<string>("cake");
-            var itemInfo = Observable.Bind(amount, itemName, (count, name) => $"{name} {count}");
-            var disposable = itemInfo.Add(info => Assert.AreEqual("apple 100", info));
+            var itemInfo = amount.CombineLatest(itemName, (count, name) => $"{name} {count}");
+            var disposable = itemInfo.Subscribe(info => Assert.AreEqual("apple 100", info));
             itemName.Data = "apple";
             disposable.Dispose();
-            disposable = itemInfo.Add(info => Assert.AreEqual("apple 5", info));
+            disposable = itemInfo.Subscribe(info => Assert.AreEqual("apple 5", info));
             amount.Data = 5;
             disposable.Dispose();
         }
@@ -68,42 +69,14 @@ namespace Yaga.Test.Documentation
 
             Assert.AreEqual(false, nullCalled);
             Assert.AreEqual(false, valueCalled);
-            someObservable.SetDefault();
+            someObservable.Data = Option.None<int>();
             Assert.AreEqual(true, nullCalled);
             Assert.AreEqual(false, valueCalled);
-            someObservable.Data = 5;
+            someObservable.Data = Option.Some(5);
             Assert.AreEqual(true, valueCalled);
 
             disposable.Dispose();
-            someObservable.Data = 6;
-        }
-
-        [Test]
-        public void OptionalObservableChainExample()
-        {
-            var nullCalled = false;
-            var valueCalled = false;
-            
-            var amount = new OptionalObservable<int>(100);
-            var itemName = new OptionalObservable<string>("cake");
-            var itemInfo = OptionalObservable.Bind(amount, itemName, (count, name) => $"{name} {count}");
-
-            var disposable = itemInfo.Subscribe(info =>
-            {
-                valueCalled = true;
-                Assert.AreEqual("apple 100", info);
-            }, () => nullCalled = true);
-            
-            Assert.AreEqual(false, nullCalled);
-            Assert.AreEqual(false, valueCalled);
-            itemName.Data = "apple";
-            Assert.AreEqual(false, nullCalled);
-            Assert.AreEqual(true, valueCalled);
-            amount.SetDefault();
-            Assert.AreEqual(true, nullCalled);
-
-            disposable.Dispose();
-            amount.Data = 6;
+            someObservable.Data = Option.Some(6);
         }
     }
 }